+++ .env.example
@@ -1,0 +1,9 @@
+NODE_ENV = 
+PORT = 
+AWS_REGION = 
+
+AWS_ACCESS_KEY_ID = ""
+AWS_SECRET_ACCESS_KEY = ""
+AWS_DEFAULT_REGION = ""
+COGNITO_USER_POOL_ID = ""
+APP_CLIENT_ID = ""
